pyproject.toml
[tool.poetry]
name = "pathway-generator"
version = "0.1.0"
description = "Generate pathways based on students background"
authors = ["Roy<roysnehendu1029@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.22.4"}
load-dotenv = "^0.1.0"
crewai-tools = "^0.0.15"
flask = "^3.0.2"
flask-cors = "^4.0.0"
langchain-google-genai = "^1.0.3"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"